services/worker-railway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg[binary]==3.2.5
python-dotenv==1.0.0
